Move keypair details modal back and move keypair create modal to the folder
