+++ src/components/Cluster/ClusterDetail/KeyPairs.js
@@ -14,8 +14,8 @@ import Copyable from 'shared/Copyable';
 import Button from 'UI/Button';
 
 import CertificateOrgsLabel from './CertificateOrgsLabel';
-import KeypairCreateModal from './KeyPairCreateModal';
-import KeyPairDetailsModal from './KeypairDetailsModal/KeyPairDetailsModal';
+import KeypairCreateModal from './KeypairCreateModal/KeyPairCreateModal';
+import KeyPairDetailsModal from './KeyPairDetailsModal';
 
 const Disclaimer = styled.p`
   margin: 0 0 20px;
